Refactor init.sh
Some major changes: Remove confirm(), and use prompt() for everything.
Prompt() now also has extra options to set the default response.

Get() now only downloads the file if it doesn't already exist.

Added replace(), to download a prebuilt conf file and replace the local
conf file.

Moved IP address steps to before IP tables.

SSH is now a separate item.

The iptables script is now a systemd unit as well.

A number of file permissions are set just in case.

Splunk installation has been moved to the end of the script since it can
take a while and is non-essential.